packages/agentcore-common/pyproject.toml
# AgentCore Common - Shared Utilities Library
# Authentication, configuration, and observability helpers

[project]
name = "agentcore-common"
version = "0.1.0"
description = "Common utilities for Amazon Bedrock AgentCore agents"
requires-python = ">=3.13"

dependencies = [
    "boto3>=1.35.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    "aws-xray-sdk>=2.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
